Implement DRS: Dynamic Resolution Switch
The Dynamic Resolution Switch (DRS) functionality allows to
change the display resolution on-the-fly.
Currently, this is done for Android N, which doesn't fully
support the DRS functionality, so this is done through a
final SurfaceFlinger restart, though the correct implementation
is already done and 99% used already.

This commit requires HAL support: even if the current one, as
said, does not support DRS completely, some bits need to get
executed for this feature to work.
The correct implementation needs the full HWC2 display HAL
implementation to work.

Thanks to @oshmoun @myself5 for Java assistance :)