Add special handling for Refresh Rate change only
When changing only the refresh rate, at least on Android N,
we can't go by the DRS path because it just checks the
height and width.... and this makes it to not switch the FPS.

Use pure sysfs for now.
When the HWC2 implementation will be complete, this will be
done in the proper way... hopefully.