Implement DispCal: Display Whitepoint Calibration
Recently I wrote a PCC Profiling facility in the somc_panel
kernel driver for SoMC devices displays.
The PCC profiling facility allows a power user, or a system
application to apply various predefined calibrations to the
device's display.

This implements the usage of this new kernel functionality
for the end user through an easy menu to choose between
all the available predefined whitepoint calibrations.