DispCal: Detect device encryption status and delay auto setting
If the device is encrypted, we cannot read the required
property to get the current DispCal profile and we cannot
write to sysfs for some reason.
Detect device encryption status and, in case the device is
encrypted, delay setting the display calibration to when
the user has unlocked the device.