Glove mode: Simplify preference update.
Instead of reverting the preference value after the fact, update it
only when the user wants it to be on, and when it succeeds applying.

At the same time, this removes the jump in the background where the
switch visually moves to the other state, and back again when the user
cancels the dialog.

Change-Id: Ice324c9174e0355f670c6d2465dc96a0ff0222de
Signed-off-by: MarijnS95 <marijns95@gmail.com>