adb-tcp: Print addresses for all adapters instead of erroring about wifi
[Especially with the addition of concurrent STA/AP] It is not uncommon
to attempt an adb connection to a phone that is hosting a WiFi hotspot,
with ExtendedSettings currently erroneously printing an error when WiFi
is disabled.
ADB over TCP is activated properly and the device is reachable on all
interfaces, thus print all their addresses. There are currently way too
many, so we might have to find a way to only print interesting ones such
as the WiFi address as previously emitted together with
isSiteLocalAddress which covers at least the hotspot.