Delete stale DRS / DispCal setup
The code for Dynamic Resolution Setting / Display Calibration is stale,
no longer supported / working and would need to be completely refactored
to work on modern kernels _and compile with modern versions of Android_
by properly talking to HWC2.

Remove it for now until someone finds the time to make it work properly
again, by reverting this commit and investigating the necessary sysfs /
HWC2 toggles to poke.